+++ tests/regression_tests_data.py
@@ -590,7 +590,7 @@ END ANSWER
 		# Test version             : 1
 		# Test since               : 2019/06/04
 		# Test last modified       : 2019/06/04
-		# Test for PolyOp version  : 1.0
+		# Test for PolyOp version  : 1.1
 		'purpose'    : 'Test the zonepred operation',
 		'input_files': ['zonepred.polyop'],
 		'options'    : '',
@@ -720,7 +720,7 @@ END ANSWER
 		# Test version             : 1
 		# Test since               : 2019/06/18
 		# Test last modified       : 2019/06/18
-		# Test for PolyOp version  : 1.0
+		# Test for PolyOp version  : 1.2
 		'purpose'    : 'Test the update operation',
 		'input_files': ['updates.polyop'],
 		'options'    : '',
@@ -841,11 +841,153 @@ BEGIN ANSWER
 & z > x
 & y = 2046
 END ANSWER
+
+
+(*--------------------*)
+
+(* OPERATION 12: 
+update ((y := x + 0, x := y + 0)) in (  y = 2
+& x = 1)
+*)
+BEGIN ANSWER
+  y = 1
+& x = 2
+END ANSWER
+
+(*--------------------*)
+
+(* OPERATION 13: 
+update ((y := 2*y + x + 5, x := y + 0)) in (  y = 2
+& x = 1)
+*)
+BEGIN ANSWER
+y = 10
+& x = 2
+END ANSWER
+
+
 			"""
 			} #end statespace file
 		] # end expectations
 	} # end test case
 			
 	,
-		### THE END
+	
+	#------------------------------------------------------------
+	{
+		# Test version             : 1
+		# Test since               : 2019/06/18
+		# Test last modified       : 2019/06/18
+		# Test for PolyOp version  : 1.2
+		'purpose'    : 'Test the zonepredgu operation',
+		'input_files': ['zonepredgr.polyop'],
+		'options'    : '',
+		'expectations' : [
+			{'file': 'zonepredgr.polyop.res' , 'content' : """
+(* OPERATION 1: 
+zonepredgu (  x = 0, true, (),   x = 0, , true, (),   x = 0)
+*)
+BEGIN ANSWER
+  x = 0
+END ANSWER
+
+
+(*--------------------*)
+
+(* OPERATION 2: 
+zonepredgu (  x = 0, true, (x := 0),   x = 0, , true, (x := 0),   x = 0)
+*)
+BEGIN ANSWER
+  x = 0
+END ANSWER
+
+
+(*--------------------*)
+
+(* OPERATION 3: 
+zonepredgu (  x = 0, true, (x := x + 0),   x = 0, , true, (x := x + 0),   x = 0)
+*)
+BEGIN ANSWER
+  x = 0
+END ANSWER
+
+
+(*--------------------*)
+
+(* OPERATION 4: 
+zonepredgu (  x >= 0,   x = 3, (x := 1),   x >= 1, x, true, (),   x = 2)
+*)
+BEGIN ANSWER
+  x = 1
+END ANSWER
+
+
+(*--------------------*)
+
+(* OPERATION 5: 
+zonepredgu (  x >= 0
+& x = y,   x = 3, (x := 1),   x >= 1
+& x + 2 = y, x, y, true, (),   x = 2
+& y = 4)
+*)
+BEGIN ANSWER
+  x = 1
+& y = 3
+END ANSWER
+
+
+(*--------------------*)
+
+(* OPERATION 6: 
+zonepredgu (  x >= 0,   x = 3, (),   x >= 1, , true, (),   x = 2)
+*)
+BEGIN ANSWER
+false
+END ANSWER
+
+
+(*--------------------*)
+
+(* OPERATION 7: 
+zonepredgu (  x = 0,   x = 0, (x := 0),   x >= 0
+& 1 >= x, x,   x = 1, (x := 0),   x = 0)
+*)
+BEGIN ANSWER
+  x = 0
+END ANSWER
+
+
+(*--------------------*)
+
+(* OPERATION 8: 
+zonepredgu (  x >= 0
+& x = y,   x = 3, (x := 1),   x >= 1
+& x + 2 = y, x, y,   y = 5, (y := 0),   x = 3
+& y = 0)
+*)
+BEGIN ANSWER
+  x = 1
+& y = 3
+END ANSWER
+
+
+(*--------------------*)
+
+(* OPERATION 9: 
+zonepredgu (  x >= 0
+& y >= 0,   x = 3, (x := 1),   x >= 1
+& x + 2 = y, x, y,   y = 5, (y := 0),   x = 3
+& y = 0)
+*)
+BEGIN ANSWER
+  x = 1
+& y = 3
+END ANSWER
+			"""
+			} #end statespace file
+		] # end expectations
+	} # end test case
+			
+	,
+			### THE END
 ]